add real user check

+++ cpas/src/main/java/com/cpas/business/concrete/PatientManager.java
@@ -2,11 +2,14 @@ package com.cpas.business.concrete;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
+import com.cpas.adapter.abstracts.UserCheckService;
 import com.cpas.business.abstracts.PatientService;
 import com.cpas.core.ExceptionHandler.UserNotFoundException;
 import com.cpas.core.utilities.results.DataResult;
+import com.cpas.core.utilities.results.ErrorResult;
 import com.cpas.core.utilities.results.Result;
 import com.cpas.core.utilities.results.SuccessDataResult;
 import com.cpas.core.utilities.results.SuccessResult;
@@ -17,15 +20,19 @@ import com.cpas.repository.PatientRepository;
 public class PatientManager implements PatientService{
 	
 	private final PatientRepository patientRepository;
+	private final UserCheckService checkService;
 	
-	public PatientManager(PatientRepository patientRepository) {
+	public PatientManager(PatientRepository patientRepository, @Qualifier("mernis") UserCheckService checkService) {
 		super();
 		this.patientRepository = patientRepository;
+		this.checkService = checkService;
 	}
 
 
 	@Override
 	public Result add(Patient patient) {
+		if(this.patientRepository.existsByNationalId(patient.getNationalId())) {return new ErrorResult("User already exists");}
+		if(!checkService.checkIfRealPerson(patient)) {return new ErrorResult("Not a valid person");}
 		this.patientRepository.save(patient);
 		return new SuccessResult("User added.");
 	}
@@ -33,6 +40,7 @@ public class PatientManager implements PatientService{
 
 	@Override
 	public Result delete(int id) {
+		if(this.patientRepository.existsById(id)) {return new ErrorResult("User not found");}
 		this.patientRepository.deleteById(id);
 		return new SuccessResult("User deleted.");
 	}
@@ -55,6 +63,4 @@ public class PatientManager implements PatientService{
 	public Patient getById(int id) throws UserNotFoundException {
 		return this.patientRepository.findById(id).orElseThrow(() -> new UserNotFoundException("User not found"));
 	}
-
-	
 }
